refactor: Remove __mocks__ imports in local-storage package
