--- packages/core/local-storage/tests/__mocks__/Logger.ts
@@ -1,14 +1,0 @@
-import { Logger } from '@verdaccio/types';
-
-const logger: Logger = {
-  warn: jest.fn(),
-  error: jest.fn(),
-  // fatal: jest.fn(),
-  info: jest.fn(),
-  debug: jest.fn(),
-  child: jest.fn(),
-  http: jest.fn(),
-  trace: jest.fn(),
-};
-
-export default logger;
